# chore(provisioner/terraform): make testdata generate.sh parallel